Dev to Master

See merge request interlay/polkabtc-clients!65

+++ staked-relayer/src/status.rs
@@ -260,7 +260,7 @@ mod tests {
                 sat: u64,
                 redeem_id: &[u8; 32],
                 op_timeout: Duration,
-                num_confirmations: u16,
+                num_confirmations: u32,
             ) -> Result<TransactionMetadata, BitcoinError>;
         }
     }
